Reorganize subpackages     - Move Areas into separate subpackage     - Move Splinter based Fields into separate module
