+++ stere/areas/__init__.py
@@ -1,0 +1,2 @@
+from .area import Area  # NOQA: F401
+from .repeating_area import RepeatingArea  # NOQA: F401
