+++ stere/fields/__init__.py
@@ -1,11 +1,9 @@
-from .fields import Field  # NOQA: F401
-from .fields import Button  # NOQA: F401
-from .fields import Input  # NOQA: F401
-from .fields import Link  # NOQA: F401
-from .fields import Root  # NOQA: F401
-from .fields import Text  # NOQA: F401
+from .field import Field  # NOQA: F401
+
+from .splinter_fields import Button  # NOQA: F401
+from .splinter_fields import Input  # NOQA: F401
+from .splinter_fields import Link  # NOQA: F401
+from .splinter_fields import Root  # NOQA: F401
+from .splinter_fields import Text  # NOQA: F401
 
 from .dropdown import Dropdown  # NOQA: F401
-
-from .area import Area  # NOQA: F401
-from .repeating_area import RepeatingArea  # NOQA: F401
