+++ stere/fields/field.py
@@ -61,42 +61,3 @@ class Field:
             List
         """
         return self._element.find()
-
-
-class Button(Field):
-    """Convenience Class on top of Field."""
-    def perform(self, value=None):
-        self.find().click()
-        return False
-
-
-class Input(Field):
-    """Convenience Class on top of Field."""
-    def perform(self, value=None):
-        self.find().fill(value)
-        return True
-
-
-class Link(Field):
-    """Convenience Class on top of Field."""
-    @property
-    def text(self):
-        elem = self.find()
-        return elem.text
-
-    def perform(self, value=None):
-        self.find().click()
-        return False
-
-
-class Root(Field):
-    def find(self):
-        return self._element.find()
-
-
-class Text(Field):
-    """Convenience Class on top of Field."""
-    @property
-    def text(self):
-        elem = self.find()
-        return elem.text
